+++ src/index.tsx
@@ -1,8 +1,9 @@
 import React from "react";
 import ReactDOM from "react-dom";
 import { BrowserRouter } from "react-router-dom";
+import App from "./App";
 import "./index.css";
-import MapPage from "./MapPage";
+import MapPage from "./pages/MapPage";
 
 ReactDOM.render(
   <React.StrictMode>
